pyproject.toml
[tool.poetry]
name = "msband"
version = "3.0.0"
description = "A library to interface with the discontinued Microsoft Band and Microsoft Band 2"
authors = [
    "‮<ku.taram@erih> nerB taraM‭"
]

license = "GOODIDEA2"
readme = "README.md"

packages = [
    { include = "msband", from = "src" }
]

include = [
    { path = "examples" },
    "LICENSE",
    "README.md",
    "pyproject.toml",
]

repository = "https://github.com/hire-marat/msband"

[tool.poetry.dependencies]
python = "^3.10"  # for dataclasses kwonly

#bleak = "^0.14"  # for connecting via BLE
construct = "*"  # for packing and unpacking wire formats
construct-typing = "0.5.2"  # typehints and objects for the wire formats
pillow = "^9.1"  # for image processing
pyusb = "^1.2"  # for interfacing via USB

[tool.poetry.dev-dependencies]  # development assistance
ipython = "*"  # nice console + PyCharm integration
poetry = "*"  # for ease of updating
pyperclip = "*"  # for ease of copy/paste
python-pcapng = "*"  # for parsing Wireshark sessions

[tool.poetry.dev-dependencies.black]  # for code formatting
version = "*"
extras = ["d"]

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core>=1.0.7"]
build-backend = "poetry.core.masonry.api"
